requirements.txt
# Core dependencies
numpy==2.0.2
pandas==2.2.3
pydicom==2.4.4
PyYAML==6.0.2

# DICOM networking and PACS interaction
pynetdicom==2.0.2

# Excel file handling
openpyxl==3.1.5
et_xmlfile==2.0.0

# Web framework and related
django>=4.2.22
django-tailwind==3.6.0
asgiref==3.8.1
sqlparse==0.5.1

# Authentication and Security
bcrypt==4.2.1
cryptography==44.0.2
pycparser==2.22

# Utilities and Parsing
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
dotenv==0.9.9
lark==1.2.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2

# Networking and Data Formats
requests>=2.32.4
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10

# Packaging and Distribution
altgraph==0.17.4
macholib==1.16.3
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
packaging==24.2
importlib_metadata==8.5.0
zipp==3.21.0

# YAML Processing
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12 